requirements.txt
flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
werkzeug==2.3.7
